+++ edaflow/__init__.py
@@ -2,7 +2,7 @@
 edaflow - A Python package for exploratory data analysis workflows
 """
 
-from .analysis import check_null_columns, analyze_categorical_columns, convert_to_numeric, visualize_categorical_values, display_column_types, impute_numerical_median, impute_categorical_mode
+from .analysis import check_null_columns, analyze_categorical_columns, convert_to_numeric, visualize_categorical_values, display_column_types, impute_numerical_median, impute_categorical_mode, visualize_numerical_boxplots
 
 __version__ = "0.4.0"
 __author__ = "Evan Low"